@@ -1,9 +1,12 @@
 package edu.tamu.richardcouperthwaite.writinglog.Repositories;
 
+
 import android.app.Application;
-import android.arch.lifecycle.LiveData;
+
+import androidx.lifecycle.LiveData;
 import android.os.AsyncTask;
 import java.util.List;
+
 import edu.tamu.richardcouperthwaite.writinglog.Database.WritingLogDatabase;
 import edu.tamu.richardcouperthwaite.writinglog.interfaces.ProjectDao;
 import edu.tamu.richardcouperthwaite.writinglog.models.Project;
