@@ -1,11 +1,11 @@
 package edu.tamu.richardcouperthwaite.writinglog.interfaces;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Delete;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
-import android.arch.persistence.room.Update;
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Delete;
+import androidx.room.Insert;
+import androidx.room.Query;
+import androidx.room.Update;
 
 import java.util.List;
 
@@ -13,7 +13,6 @@ import edu.tamu.richardcouperthwaite.writinglog.models.Project;
 
 @Dao
 public interface ProjectDao {
-
     @Insert
     void insert(Project project);
 
@@ -28,5 +27,4 @@ public interface ProjectDao {
 
     @Delete
     void deleteProject(Project project);
-
 }
