@@ -1,14 +1,13 @@
 package edu.tamu.richardcouperthwaite.writinglog.interfaces;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.Query;
 
 import java.util.List;
 
 import edu.tamu.richardcouperthwaite.writinglog.models.Session;
-
 
 @Dao
 public interface sessionDao {
