@@ -1,15 +1,14 @@
 package edu.tamu.richardcouperthwaite.writinglog.interfaces;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
-import android.arch.persistence.room.Update;
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.Query;
+import androidx.room.Update;
 
 import java.util.List;
 
 import edu.tamu.richardcouperthwaite.writinglog.models.Statistics;
-
 
 @Dao
 public interface statsDao {
