@@ -1,11 +1,13 @@
 package edu.tamu.richardcouperthwaite.writinglog.models;
 
 import android.app.Application;
-import android.arch.lifecycle.AndroidViewModel;
-import android.arch.lifecycle.LiveData;
-import java.util.List;
-import edu.tamu.richardcouperthwaite.writinglog.Repositories.sessionRepository;
 
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.LiveData;
+
+import java.util.List;
+
+import edu.tamu.richardcouperthwaite.writinglog.Repositories.sessionRepository;
 
 public class sessionViewModel extends AndroidViewModel {
     private sessionRepository mRepository;
